// src/app/api/generated/model/limit-usage-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Consumo actual de un limite del plan contratado
 */
export interface LimitUsageResponse { 
    /**
     * Codigo del limite
     */
    code?: LimitUsageResponseCodeEnum;
    /**
     * Cantidad de recursos activos que consumen este limite
     */
    currentUsage?: number;
    /**
     * true cuando el uso actual ya supera el tope del plan. Un downgrade puede dejarlo en true sin borrar nada: lo existente sigue operativo y solo se rechazan las altas nuevas
     */
    exceeded?: boolean;
    /**
     * Tope que impone el plan. Ausente cuando el limite es ilimitado
     */
    limitValue?: number;
}
export enum LimitUsageResponseCodeEnum {
    MAX_CONSULTORIOS = 'MAX_CONSULTORIOS',
    MAX_MIEMBROS_ACTIVOS = 'MAX_MIEMBROS_ACTIVOS'
};